requirements.txt
django
djangorestframework
django-cors-headers
mysqlclient
Pillow
pdf2image
google-generativeai